requirements.txt
transformers
Pillow
flask
Flask-Reuploaded
Flask-WTF
Werkzeug==2.2.2
torch
requests
